Make private column._element property writable, check before defining
Calling Object.defineProperty(...) in px-data-grid-column#_mappedObjectChanged
was created a non-writable property on the column object. Sometimes
when we hit that method again for a column that had already been
used a TypeError was thrown saying that _element was non-writable and
could not be re-assigned using Object.defineProperty().

Adds a check to ensure that _element should be redefined if we hit
the method again, and marks the _element property as explicitly
writable when it is defined to ensure we don't hit the runtime
error again when redefining it.